convert privacy policy page to tsx, use RemoteMarkdown component
